requirements.txt
flask==3.0.0
line-bot-sdk==3.7.0
azure-ai-vision-imageanalysis==1.0.0b1
groq==0.4.2
supabase==2.3.0
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.1.0
gunicorn==21.2.0

